% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsiv-data.R
\docType{data}
\name{CoRSIV_Probes}
\alias{CoRSIV_Probes}
\title{This dataframe contains hm450k probes overlapping CoRSIV regions}
\format{A data frame with 53940 rows and 10 variables}
\usage{
data(CoRSIV_Probes)
}
\description{
A dataset containing the probe id and overlapping CoRSIV ID
}
\keyword{datasets}
