% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsiv-data.R
\docType{data}
\name{CoRSIV_Probes}
\alias{CoRSIV_Probes}
\title{Prices of 50,000 round cut diamonds}
\format{A data frame with 53940 rows and 10 variables}
\usage{
data(CoRSIV_Probes)
}
\description{
A dataset containing the prices and other attributes of almost 54,000
 diamonds. The variables are as follows:
}
\keyword{datasets}
